CRUD DE CUSTOMER Y ORDER LISTOS, CRUD DE ORDERITEM ERROR AL CARGAR METODO POST

+++ src/main/java/ecommerce/consiti/ecommerce/service/impl/OrderItemService.java
@@ -1,22 +1,51 @@
 package ecommerce.consiti.ecommerce.service.impl;
 
+import ecommerce.consiti.ecommerce.entity.Order;
 import ecommerce.consiti.ecommerce.entity.OrderItem;
+import ecommerce.consiti.ecommerce.entity.Product;
+import ecommerce.consiti.ecommerce.repository.ICustomerRepository;
 import ecommerce.consiti.ecommerce.repository.IOrderItemRepository;
+import ecommerce.consiti.ecommerce.repository.IOrderRepository;
+import ecommerce.consiti.ecommerce.repository.IProductRepository;
 import ecommerce.consiti.ecommerce.service.IOrdenItemService;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
+
 @Service
 public class OrderItemService implements IOrdenItemService {
-    private IOrderItemRepository orderItemRepository;
+
+    private final IOrderItemRepository orderItemRepository;
 
     public OrderItemService(IOrderItemRepository orderItemRepository) {
         this.orderItemRepository = orderItemRepository;
     }
 
+    @Autowired
+    private IOrderRepository orderRepository;
+    @Autowired
+    private IProductRepository productRepository;
+
     @Override
     public OrderItem createOrderItem(OrderItem orderItem) {
+        Optional<Product> productOptional = productRepository.findById(orderItem.getProduct().getProductId());
+        if (!productOptional.isPresent()) {
+            throw new IllegalArgumentException("El producto no existe");
+        }
+        Product product = productOptional.get();
+        if (orderItem.getQuantity() == null || orderItem.getQuantity() < 1) {
+            throw new IllegalArgumentException("La cantidad debe ser mayor que 0");
+        }
+        orderItem.setPreci(product.getPrice() * orderItem.getQuantity());
+        Optional<Order> orderOptional = orderRepository.findById(orderItem.getOrder().getOrderId());
+        if (!orderOptional.isPresent()) {
+            throw new IllegalArgumentException("La orden no existe");
+        }
+        Order order = orderOptional.get();
+        order.setTotalAmount(order.getTotalAmount() + orderItem.getPreci());
+        orderItem.setOrder(order);
         return orderItemRepository.save(orderItem);
     }
 
@@ -29,14 +58,47 @@ public class OrderItemService implements IOrdenItemService {
     public List<OrderItem> searchAll() {
         return orderItemRepository.findAll();
     }
-
     @Override
-    public void updateOrderItem(OrderItem orderItem) {
-        orderItemRepository.save(orderItem);
+    public boolean updateOrderItem(Integer id, OrderItem orderItem) {
+        Optional<OrderItem> orderItemOptional = orderItemRepository.findById(id);
+        if (orderItemOptional.isPresent()) {
+            OrderItem existingOrderItem = orderItemOptional.get();
+
+            if (orderItem.getProduct() != null) {
+                Optional<Product> productOptional = productRepository.findById(orderItem.getProduct().getProductId());
+                if (!productOptional.isPresent()) {
+                    throw new IllegalArgumentException("El producto no existe");
+                }
+                existingOrderItem.setProduct(productOptional.get());
+            }
+
+            if (orderItem.getQuantity() != null && orderItem.getQuantity() >= 1) {
+                existingOrderItem.setQuantity(orderItem.getQuantity());
+            } else {
+                throw new IllegalArgumentException("La cantidad debe ser mayor que 0");
+            }
+
+            existingOrderItem.setPreci(existingOrderItem.getProduct().getPrice() * existingOrderItem.getQuantity());
+            if (existingOrderItem.getOrder() != null) {
+                existingOrderItem.getOrder().setTotalAmount(existingOrderItem.getOrder().getTotalAmount() + existingOrderItem.getPreci());
+            }
+
+            orderItemRepository.save(existingOrderItem);
+            return true;
+        }
+        return false;
     }
+
 
     @Override
     public void deleteOrderItem(Integer id) {
-        orderItemRepository.deleteById(id);
+        Optional<OrderItem> orderItemOptional = orderItemRepository.findById(id);
+        if (orderItemOptional.isPresent()) {
+            OrderItem orderItem = orderItemOptional.get();
+            if (orderItem.getOrder() != null) {
+                orderItem.getOrder().setTotalAmount(orderItem.getOrder().getTotalAmount() - orderItem.getPreci());
+            }
+            orderItemRepository.deleteById(id);
+        }
     }
 }
